Broadcast incoming transaction over websocket

+++ src/main/java/io/smilo/api/peer/payloadhandler/sport/NetworkState.java
@@ -88,7 +88,7 @@ public class NetworkState implements INetworkState {
         return topBlock;
     }
 
-    public void setTopBlock(long topBlock) {
+    public void setTopBlock(int topBlock) {
         this.topBlock = topBlock;
         updateCatchupMode();
     }
